Fixes tests
No where is using the ecdsa schemes other than to yell at you if your
curve size doesn't match the scheme you use... Never mind the fact that
the code then completely disregards whatever scheme is assigned to the
key and uses the key's curve size to figure out how to sign the data.

Honestly, this whole function should either be removed or the scheme
actually used. I'm in favor of removing the code, but for now I just
want the tests to pass again.