46 - bring rootcas and intermediatecas layout field to ite7 spec
ITE7 dictates that the rootcas and intermediatecas fields within the
layout should resemble the keys field.  Basically instead of being an
array of strings they should be objects indexed by the key's id with
each value being the in-toto key object.

Initially I wanted to put the certificate in the public field of keyval
but this proved challenging when trying to verify and create signatures.
So instead I chose to add a certificate field to each keyval.  This has
the unfortunate byproduct of changing every key's key id.